ALLES WERKT, AFBLIJVEN MET DE POTEN!

+++ lib/Display/Display.cpp
@@ -62,7 +62,7 @@ void Display::displayCharacters() {
 void Display::aftellen(char tienden, char enen) {
     while (true) {
         setCharacters(tienden, enen);
-        int startMillis = millis();
+        long int startMillis = millis();
         while(millis() - startMillis < 1000) {
             displayCharacters();
         }
@@ -76,7 +76,7 @@ void Display::aftellen(char tienden, char enen) {
             enen = enen - 1;
         } 
     }
-    int startMillis = millis();
+    long int startMillis = millis();
     while(millis() - startMillis < 1000) {
         setCharacters('S', 'T');
         displayCharacters();
@@ -96,18 +96,18 @@ void Display::displayFinishTime() {
     int enen = getLastTwoDigitsChar(seconds, 1);
 
     for (int i = 3; i > 0; i--) {
-        int startMillis = millis();
+        long int startMillis = millis();
         while(millis() - startMillis < 1000) {
             setCharacters(tienden, enen);
             displayCharacters();
         }
-        int startMillis2 = millis();
+        long int startMillis2 = millis();
         while(millis() - startMillis2 < 1000) {
             setCharacters(' ', ' ');
             displayCharacters();
         }
     }
-    int startMillis3 = millis();
+    long int startMillis3 = millis();
     while(millis() - startMillis3 < 1000) {
         setCharacters('F', 'I');
         displayCharacters();
